Update Manjaro11thgen.md
### Append the following to the GRUB_CMDLINE_LINUX_DEFAULT="quiet splash section.
If encrypted, make sure you place this correctly or use the copy/paste command above instead.

This is an ACPI parameter that helps ensure compatibility by disabling ACPI support for NVMe. Advanced Linux users: You're welcome to remove it if you feel it's not needed for any reason and sudo update-grub.