Cleanup and simplify ubuntu-22.04-fw-16.md tutorial
The intention is to simplify and correct the presentation of
this tutorial, not necessarily to correct or alter the commands and
steps themselves. There are still plenty of commands and steps that
seem slightly off target, but they don't appear to be entirely harmful,
as written.

It is still far from perfect, but at least it is easier for power users
to interpret the commands presented and reverse engineer the intention.