Bring in required Convolver impact file for FW13 easyeffects.
This irs file is required for the convolver plugin to work,
which creates the most significant change on the sound quality
of all the other settings. Without this file, the convolver plugin
errors out with 'Failed to load Impact file' and does not work.
With this file, the sound sounds great!

File sourced from https://github.com/cab404/framework-dsp/tree/master/config/irs