Update readme.md
NOTE: This may not disable the keyboard backlighting when you place it into suspend. By default without this script, the keyboard backlight goes out automatically. With this script,you will need to Fn space bar to turn off the backlight before you enter suspend or it may remain on. This is a side effect of the script.